spec page and organize
