+++ src/components/LessonPage/FITBBlank.tsx
