+++ src/components/LessonPage/HintDisplay.tsx
