+++ src/components/SolvePage/ClickOutside.tsx
